pyadl/requirements.txt
webhdfs>=0.4.1
